# By Ralph Samer (19) and others
# Via Ralph Samer (4) and others
* 'master' of gh:Catrobat/Catty: (25 commits)
  small changes for audio handling
  Scheme shared
  replace NSLog with NSDebug in brick-actions
  fix wrong app name
  fixes #361
  fix SearchStoreViewController storyboard issue
  remove UISearchController
  separate development targets for iOS7.1 and iOS8.0
  fix ProgramTVCExistingProgramsTests for iOS8
  minor dealloc improvements
  fixes iOS8 program bug
  move brick-object references to parser
  fix compile error
  fix merge bug
  continued on fixing iOS8 issues
  fix analyze issues + some iOS8 deprecated issues
  fixes #220
  refactor all localized-string-constants
  ad ios8_fixes
  deployment target fix
  ...

Conflicts:
	src/Catty.xcodeproj/project.pbxproj
	src/Catty/App-Prefix.pch
	src/Catty/AppDefines.h
	src/Catty/BaseTableViewController.m
	src/Catty/BrickDetailViewController.m
	src/Catty/CatrobatTableViewController.m
	src/Catty/CreateView.m
	src/Catty/FileManager.m
	src/Catty/LanguageTranslationDefines.h
	src/Catty/LooksTableViewController.m
	src/Catty/MyProgramsViewController.m
	src/Catty/Program.h
	src/Catty/Program.m
	src/Catty/ProgramTableViewController.m
	src/Catty/ProjectParser.m
	src/Catty/ScriptCollectionViewController.m
	src/Catty/SearchStoreViewController.m
	src/Catty/SoundsTableViewController.m

+++ src/Catty/SMPageControl.m
@@ -419,8 +419,8 @@ static SMPageControlStyleDefaults _defaultStyleForSystemVersion;
 
 - (CGImageRef)createMaskForImage:(UIImage *)image CF_RETURNS_RETAINED
 {
-	size_t pixelsWide = image.size.width * image.scale;
-	size_t pixelsHigh = image.size.height * image.scale;
+	size_t pixelsWide = (size_t)(image.size.width * image.scale);
+	size_t pixelsHigh = (size_t)(image.size.height * image.scale);
 	size_t bitmapBytesPerRow = (pixelsWide * 1);
 	CGContextRef context = CGBitmapContextCreate(NULL, pixelsWide, pixelsHigh, CGImageGetBitsPerComponent(image.CGImage), bitmapBytesPerRow, NULL, (CGBitmapInfo)kCGImageAlphaOnly);
 	CGContextTranslateCTM(context, 0.f, pixelsHigh);
